requirements.txt
python-dotenv==1.0.0
googletrans==4.0.0rc1
llama-index==0.6.9
scikit-learn==1.2.2
rich==13.3.5
langchain==0.0.177
python-dotenv==1.0.0
nltk==3.8.1
tqdm==4.65.0
rank-bm25==0.2.2
pypdf==3.9.0
